% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{datasets}
\alias{datasets}
\title{mydas datasets}
\description{
Example datasets for the classes defined in FLCore.
}

See [mydas_turbot](https://3o2y9wugzp1kfxr5hvzgzq-on.drv.tw/MyDas/om/mydas_turbot.html) vignette for more infomation

Datasets can be loaded by issuing the \code{data} command, as in
\code{data(om)}
\code{data(eq)}
}
\examples{
\dontrun{
data(om)
summary(om)
}
}
\seealso{
\linkS4class{FLStock}, \linkS4class{FLBRP}, \linkS4class{FLPar}
}
\keyword{datasets}
